+++ src/heijunka/StateHistory.ts
@@ -20,7 +20,15 @@ export class StateHistory {
         }
         const newTransitions = [...this.transitions];
         newTransitions.push(aNewTransition);
-        newTransitions.sort((a,b)=> a.occurredAt.valueOf() - b.occurredAt.valueOf() );
+        newTransitions.sort((a, b) => a.occurredAt.valueOf() - b.occurredAt.valueOf());
         return new StateHistory(newTransitions);
+    }
+
+    currentState(): string {
+        if (this.transitions.length === 0) {
+            throw new Error('not attached to a state at the moment.');
+        } else {
+            return this.transitions[0].state;
+        }
     }
 }