+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.2.0
+- Boilerplate code for finding cards and projects based on their id
+- Encapsulate finding initial state as well
+
 # v1.1.0
 - Events available for creating a Kanban Card and moving it around